requirements.txt
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
